pyproject.toml
[tool.poetry]
name = "product-api"
version = "0.1.0"
description = "API ordering the products"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.27.2"
config = "^0.5.1"
factory-boy = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
